Submit payment done

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -64,9 +64,29 @@ export class CheckoutPaymentComponent implements OnInit {
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: order => {
         this.toastr.success('Order created successfully');
-        this.basketService.deleteLocalBasket();
-        const navigationExtras: NavigationExtras = {state: order};
-        this.router.navigate(['checkout/success'], navigationExtras);
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value,
+              address: {
+                city: 'Bengaluru',
+                country: 'IN',
+                line1: '#123, KR Layout, 3rd block',
+                line2: 'JP Nagar 6th Phase',
+                postal_code: '560078',
+                state: 'Karnataka'
+              }
+            }
+          }
+        }).then(result => {
+          console.log(result);
+          if(result.paymentIntent) {
+            this.basketService.deleteLocalBasket();
+            const navigationExtras: NavigationExtras = {state: order};
+            this.router.navigate(['checkout/success'], navigationExtras);
+          }
+        })
       }
     })
   }
